[Slide]: Performance improvements and slot props support
Converted all of the isX values to computed refs. This means that the callbacks are only re-executed when dependent values change rather than every time the render function runs.

Also added all the isX values to the slotProps provided to slide. This can provide many useful functions to the end-user, including being able to respond to changes in visibility when using their own ARIA props inside the slot.